add testing

+++ internal/handler/model.go
@@ -2,6 +2,6 @@ package handler
 
 // CreateProductRequest Структура для входящих данных продукта
 type CreateProductRequest struct {
-	Name  string  `json:"name" binding:"required"`
-	Price float64 `json:"price" binding:"required"`
+	Name  string `json:"name" binding:"required"`
+	Price int    `json:"price" binding:"required"`
 }
